Enhancement  | Consider glossary words occurrences
See: 353

This counts how many times a glossary original word is used in the original and then when checking the translation, it validates the translation if the translation contains the suggested translation at least the same number of times as in original.

This DOES NOT fix the issue with validation glossary of the plural strings.