Check for glossary words all translation forms
Fixes: #359

- Defines SINGULAR and PLURAL constants to easier understand the code later.
- Uses jQuery as to keep the old logic.
- Queries all translations (singular and plural/plurals if any) and all originals (singular and plural)

First, it will compare with original singular, but if the locale has `nplural===1` (for e.g. ja) and string is plural, it will compare with original plural.

Each translation is compared with the corresponding original:
- first translation with original singular
- second translation with original plural
- any other plural translation forms with original plural

Next, the algorithm stays the same, just some variable renaming, for clarity.

Finally, if the string is plural, the error will specify its origin: plural or singular form.